extend common_settings_interfaces.dart documentation-
