+++ lib/model/storage/export_csv_settings_store.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 
 import 'package:blood_pressure_app/model/export_options.dart';
-import 'package:blood_pressure_app/model/storage/common_settings_intervaces.dart';
+import 'package:blood_pressure_app/model/storage/common_settings_interfaces.dart';
 import 'package:blood_pressure_app/model/storage/convert_util.dart';
 import 'package:flutter/material.dart';
